requirements.txt
os
pillow
matplotlib
numpy
tensorflow
keras
sklearn